+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+* Add `check-when-delete-rows` util for check actions that triggered before delete row and confirm.
 
 ## [2.6.0-beta.1] - 2020-11-15
 ### Added